optimize conf process logic direct in internal/conf package
